requirements.txt
setuptools
wheel
numpy>=1.21.0
scipy>=1.7.0
torch>=1.9.0
torchvision>=0.10.0
matplotlib>=3.4.0
opencv-python>=4.5.0
Pillow>=8.0.0
tqdm>=4.60.0
gym>=0.18.0
atari-py>=0.2.9
PyYAML>=5.4.0
omegaconf>=2.1.0
tensorboard>=2.6.0
scikit-learn>=0.24.0
pandas>=1.3.0
tensorboardX>=2.3
h5py
hydra-core
cma
swig==4.4.0
minari==0.5.2
minari[hf]
gymnasium[box2d]
gymnasium[atari]